unikontainers: Refactor UnikernelConfig.decode and getBlockDevice to enable unit tests
- refactor UnikernelConfig.decode function to return error instead of exiting
- refactor getBlockDevice to accept a getter function to enable mocking in tests
- add missing copyright comment

PR: https://github.com/nubificus/urunc/pull/42
Approved-by: Charalampos Mainas <cmainas@nubificus.co.uk>
